Split injector.ex into individual files and add @moduledoc
Move each module from the monolithic injector.ex into its own file
following Elixir's one-file-per-module convention. Add @moduledoc to
Faultex.Injector, Faultex.Response, RandomInjector, and ChainInjector.
