+++ lib/faultex/injector/slow_injector.ex
@@ -1,0 +1,45 @@
+defmodule Faultex.Injector.SlowInjector do
+  @moduledoc """
+  Inject response delay
+  """
+
+  @type t :: %__MODULE__{
+          id: term(),
+          disable: boolean() | nil,
+          host: String.t() | nil,
+          method: String.t() | nil,
+          path: String.t() | nil,
+          headers: [{String.t(), String.t()}] | nil,
+          percentage: integer() | nil,
+          resp_delay: integer() | nil
+        }
+
+  defstruct [
+    :id,
+    :disable,
+    :host,
+    :method,
+    :path,
+    :headers,
+    :percentage,
+    :resp_delay
+  ]
+
+  @behaviour Faultex.Injector
+
+  @impl Faultex.Injector
+  @spec inject(t()) :: Faultex.Response.t()
+  def inject(injector) do
+    resp_delay =
+      case Map.get(injector, :resp_delay) do
+        nil -> 0
+        delay -> delay
+      end
+
+    if resp_delay != 0 do
+      Process.sleep(resp_delay)
+    end
+
+    %Faultex.Response{}
+  end
+end
